Move class to different package and change imports
