+++ core/src/integrationTest/java/tech/pegasys/web3signer/core/jsonrpcproxy/EthAccountsIntegrationTest.java
@@ -14,8 +14,8 @@ package tech.pegasys.web3signer.core.jsonrpcproxy;
 
 import static java.util.Collections.singletonList;
 
+import tech.pegasys.web3signer.core.Eth1AddressSignerIdentifier;
 import tech.pegasys.web3signer.core.service.jsonrpc.response.JsonRpcSuccessResponse;
-import tech.pegasys.web3signer.core.util.Eth1AddressSignerIdentifier;
 
 import java.util.Map.Entry;
