+++ core/src/main/java/tech/pegasys/web3signer/core/service/jsonrpc/handlers/internalresponse/EthAccountsResultProvider.java
@@ -10,7 +10,7 @@
  * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations under the License.
  */
-package tech.pegasys.web3signer.core.service.http.handlers.internalresponse;
+package tech.pegasys.web3signer.core.service.jsonrpc.handlers.internalresponse;
 
 import tech.pegasys.web3signer.core.Eth1AddressSignerIdentifier;
 import tech.pegasys.web3signer.core.service.jsonrpc.JsonRpcRequest;
