+++ core/src/test/java/tech/pegasys/web3signer/core/service/jsonrpc/EthAccountsResultProviderTest.java
@@ -19,8 +19,8 @@ import static org.assertj.core.api.Assertions.catchThrowable;
 import static tech.pegasys.web3signer.core.service.jsonrpc.response.JsonRpcError.INVALID_PARAMS;
 
 import tech.pegasys.signers.secp256k1.EthPublicKeyUtils;
-import tech.pegasys.web3signer.core.service.http.handlers.internalresponse.EthAccountsResultProvider;
 import tech.pegasys.web3signer.core.service.jsonrpc.exceptions.JsonRpcException;
+import tech.pegasys.web3signer.core.service.jsonrpc.handlers.internalresponse.EthAccountsResultProvider;
 
 import java.security.interfaces.ECPublicKey;
 import java.util.List;
